requirements.txt
# RSS/Feed parsing
feedparser>=6.0.10

# OpenAI SDK
openai>=1.14.0

# HTTP client
requests>=2.31.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Date parsing
python-dateutil>=2.8.2

# YAML config
pyyaml>=6.0.1

# Retry logic
tenacity>=8.2.3

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
